+++ tests/test_insert.cpp
@@ -1,5 +1,6 @@
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
+
 #include <bits/bits.h>
 
 using ::testing::ElementsAre;