+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Change log
 
 ## Not yet released
+- Hide implementation detail under 'bits::detail' namespace
 - Initial version